+++ src/iir/mod.rs
@@ -12,3 +12,5 @@ mod sos;
 pub use sos::*;
 mod normal;
 pub use normal::*;
+mod wdf;
+pub use wdf::*;